+++ src/main/java/frc/robot/subsystems/drive/ModuleIOKrakenFOC.java
@@ -1,0 +1,202 @@
+// Copyright 2021-2024 FRC 6328
+// http://github.com/Mechanical-Advantage
+//
+// This program is free software; you can redistribute it and/or
+// modify it under the terms of the GNU General Public License
+// version 3 as published by the Free Software Foundation or
+// available in the root directory of this project.
+//
+// This program is distributed in the hope that it will be useful,
+// but WITHOUT ANY WARRANTY; without even the implied warranty of
+// MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
+// GNU General Public License for more details.
+
+package frc.robot.subsystems.drive;
+
+import static frc.robot.subsystems.drive.DriveConstants.*;
+
+import com.ctre.phoenix6.BaseStatusSignal;
+import com.ctre.phoenix6.StatusCode;
+import com.ctre.phoenix6.StatusSignal;
+import com.ctre.phoenix6.configs.CANcoderConfiguration;
+import com.ctre.phoenix6.configs.MotorOutputConfigs;
+import com.ctre.phoenix6.configs.TalonFXConfiguration;
+import com.ctre.phoenix6.controls.VoltageOut;
+import com.ctre.phoenix6.hardware.CANcoder;
+import com.ctre.phoenix6.hardware.TalonFX;
+import com.ctre.phoenix6.signals.InvertedValue;
+import com.ctre.phoenix6.signals.NeutralModeValue;
+import edu.wpi.first.math.geometry.Rotation2d;
+import edu.wpi.first.math.util.Units;
+import frc.robot.subsystems.drive.DriveConstants.ModuleConfig;
+import java.util.Queue;
+
+/**
+ * Module IO implementation for Talon FX drive motor controller, Talon FX turn motor controller, and
+ * CANcoder
+ *
+ * <p>NOTE: This implementation should be used as a starting point and adapted to different hardware
+ * configurations (e.g. If using an analog encoder, copy from "ModuleIOSparkMax")
+ *
+ * <p>To calibrate the absolute encoder offsets, point the modules straight (such that forward
+ * motion on the drive motor will propel the robot forward) and copy the reported values from the
+ * absolute encoders using AdvantageScope. These values are logged under
+ * "/Drive/ModuleX/TurnAbsolutePositionRad"
+ */
+public class ModuleIOKrakenFOC implements ModuleIO {
+  private final TalonFX driveTalon;
+  private final TalonFX turnTalon;
+  private final CANcoder cancoder;
+
+  private final Queue<Double> timestampQueue;
+
+  private final StatusSignal<Double> drivePosition;
+  private final Queue<Double> drivePositionQueue;
+  private final StatusSignal<Double> driveVelocity;
+  private final StatusSignal<Double> driveAppliedVolts;
+  private final StatusSignal<Double> driveCurrent;
+
+  private final StatusSignal<Double> turnAbsolutePosition;
+  private final StatusSignal<Double> turnPosition;
+  private final Queue<Double> turnPositionQueue;
+  private final StatusSignal<Double> turnVelocity;
+  private final StatusSignal<Double> turnAppliedVolts;
+  private final StatusSignal<Double> turnCurrent;
+
+  private final Rotation2d absoluteEncoderOffset;
+
+  public ModuleIOKrakenFOC(ModuleConfig config) {
+    driveTalon = new TalonFX(config.driveID(), canbus);
+    turnTalon = new TalonFX(config.turnID(), canbus);
+    cancoder = new CANcoder(config.absoluteEncoderChannel(), canbus);
+    absoluteEncoderOffset = config.absoluteEncoderOffset();
+
+    var driveConfig = new TalonFXConfiguration();
+    driveConfig.CurrentLimits.StatorCurrentLimit = 40.0;
+    driveConfig.CurrentLimits.StatorCurrentLimitEnable = true;
+
+    var turnConfig = new TalonFXConfiguration();
+    turnConfig.CurrentLimits.SupplyCurrentLimit = 30.0;
+    turnConfig.CurrentLimits.SupplyCurrentLimitEnable = true;
+    turnConfig.Voltage.PeakForwardVoltage = 12.0;
+    turnConfig.Voltage.PeakReverseVoltage = -12.0;
+
+    driveConfig.Feedback.SensorToMechanismRatio = moduleConstants.driveReduction();
+    turnConfig.Feedback.SensorToMechanismRatio = moduleConstants.turnReduction();
+    turnConfig.ClosedLoopGeneral.ContinuousWrap = true;
+
+    for (int i = 0; i < 4; i++) {
+        boolean error = driveTalon.getConfigurator().apply(driveConfig, 0.1) == StatusCode.OK;
+        setDriveBrakeMode(true);
+        error = error && turnTalon.getConfigurator().apply(turnConfig, 0.1) == StatusCode.OK;
+        setTurnBrakeMode(true);
+        error = error && cancoder.getConfigurator().apply(new CANcoderConfiguration(), 0.1) == StatusCode.OK;
+        if (!error) break;
+    }
+
+    timestampQueue = PhoenixOdometryThread.getInstance().makeTimestampQueue();
+
+    drivePosition = driveTalon.getPosition();
+    drivePositionQueue =
+        PhoenixOdometryThread.getInstance().registerSignal(driveTalon, driveTalon.getPosition());
+    driveVelocity = driveTalon.getVelocity();
+    driveAppliedVolts = driveTalon.getMotorVoltage();
+    driveCurrent = driveTalon.getStatorCurrent();
+
+    turnAbsolutePosition = cancoder.getAbsolutePosition();
+    turnPosition = turnTalon.getPosition();
+    turnPositionQueue =
+        PhoenixOdometryThread.getInstance().registerSignal(turnTalon, turnTalon.getPosition());
+    turnVelocity = turnTalon.getVelocity();
+    turnAppliedVolts = turnTalon.getMotorVoltage();
+    turnCurrent = turnTalon.getStatorCurrent();
+
+    BaseStatusSignal.setUpdateFrequencyForAll(
+        odometryFrequency, drivePosition, turnPosition);
+    BaseStatusSignal.setUpdateFrequencyForAll(
+        100.0,
+        driveVelocity,
+        driveAppliedVolts,
+        driveCurrent,
+        turnAbsolutePosition,
+        turnVelocity,
+        turnAppliedVolts,
+        turnCurrent);
+    driveTalon.optimizeBusUtilization(1.0);
+    turnTalon.optimizeBusUtilization(1.0);
+  }
+
+  @Override
+  public void updateInputs(ModuleIOInputs inputs) {
+    BaseStatusSignal.refreshAll(
+        drivePosition,
+        driveVelocity,
+        driveAppliedVolts,
+        driveCurrent,
+        turnAbsolutePosition,
+        turnPosition,
+        turnVelocity,
+        turnAppliedVolts,
+        turnCurrent);
+
+    inputs.drivePositionRad =
+        Units.rotationsToRadians(drivePosition.getValueAsDouble());
+    inputs.driveVelocityRadPerSec =
+        Units.rotationsToRadians(driveVelocity.getValueAsDouble());
+    inputs.driveAppliedVolts = driveAppliedVolts.getValueAsDouble();
+    inputs.driveCurrentAmps = new double[] {driveCurrent.getValueAsDouble()};
+
+    inputs.turnAbsolutePosition =
+        Rotation2d.fromRotations(turnAbsolutePosition.getValueAsDouble())
+            .minus(absoluteEncoderOffset);
+    inputs.turnPosition =
+        Rotation2d.fromRotations(turnPosition.getValueAsDouble());
+    inputs.turnVelocityRadPerSec =
+        Units.rotationsToRadians(turnVelocity.getValueAsDouble());
+    inputs.turnAppliedVolts = turnAppliedVolts.getValueAsDouble();
+    inputs.turnCurrentAmps = new double[] {turnCurrent.getValueAsDouble()};
+
+    inputs.odometryTimestamps =
+        timestampQueue.stream().mapToDouble((Double value) -> value).toArray();
+    inputs.odometryDrivePositionsRad =
+        drivePositionQueue.stream()
+            .mapToDouble(Units::rotationsToRadians)
+            .toArray();
+    inputs.odometryTurnPositions =
+        turnPositionQueue.stream()
+            .map(Rotation2d::fromRotations)
+            .toArray(Rotation2d[]::new);
+    timestampQueue.clear();
+    drivePositionQueue.clear();
+    turnPositionQueue.clear();
+  }
+
+  @Override
+  public void setDriveVoltage(double volts) {
+    driveTalon.setControl(new VoltageOut(volts));
+  }
+
+  @Override
+  public void setTurnVoltage(double volts) {
+    turnTalon.setControl(new VoltageOut(volts));
+  }
+
+  @Override
+  public void setDriveBrakeMode(boolean enable) {
+    var config = new MotorOutputConfigs();
+    config.Inverted = InvertedValue.CounterClockwise_Positive;
+    config.NeutralMode = enable ? NeutralModeValue.Brake : NeutralModeValue.Coast;
+    driveTalon.getConfigurator().apply(config);
+  }
+
+  @Override
+  public void setTurnBrakeMode(boolean enable) {
+    var config = new MotorOutputConfigs();
+    config.Inverted =
+        DriveConstants.moduleConfigs[0].turnMotorInverted()
+            ? InvertedValue.Clockwise_Positive
+            : InvertedValue.CounterClockwise_Positive;
+    config.NeutralMode = enable ? NeutralModeValue.Brake : NeutralModeValue.Coast;
+    turnTalon.getConfigurator().apply(config);
+  }
+}
